back/pyproject.toml
[tool.poetry]
name = "back-buscador-libros"
version = "0.1.0"
description = ""
authors = ["Giovanny Baron <gio.baronc@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "5.0.7"
djangorestframework = "^3.15.2"
unidecode = "^1.3.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
